+++ apps/www/src/constants/bookmark.ts
@@ -2,3 +2,13 @@ export enum LINK_TYPE {
   SIMILARITY = "SIMILARITY",
   KEYWORD = "KEYWORD",
 }
+
+export enum GRAPH_THEME {
+  GRAPH = "GRAPH",
+  PLANET = "PLANET",
+}
+
+export enum GRAPH_TYPE {
+  COLOR = "COLOR",
+  LOGO = "LOGO",
+}
