+++ apps/www/src/lib/zustand/bookmark.ts
@@ -1,4 +1,4 @@
-import { LINK_TYPE } from "@/constants/bookmark";
+import { GRAPH_THEME, GRAPH_TYPE, LINK_TYPE } from "@/constants/bookmark";
 
 import { create } from "zustand";
 
@@ -6,10 +6,18 @@ type FilterType = LINK_TYPE.SIMILARITY | LINK_TYPE.KEYWORD | "";
 
 interface BookmarkStore {
   selectedFilter: FilterType;
+  selectedTheme: GRAPH_THEME;
+  selectedType: GRAPH_TYPE;
   setSelectedFilter: (selectedFilter: FilterType) => void;
+  setSelectedTheme: (selectedTheme: GRAPH_THEME) => void;
+  setSelectedType: (selectedType: GRAPH_TYPE) => void;
 }
 
 export const useBookmarkStore = create<BookmarkStore>((set) => ({
   selectedFilter: "",
+  selectedTheme: GRAPH_THEME.PLANET,
+  selectedType: GRAPH_TYPE.COLOR,
   setSelectedFilter: (selectedFilter: FilterType) => set({ selectedFilter }),
+  setSelectedTheme: (selectedTheme: GRAPH_THEME) => set({ selectedTheme }),
+  setSelectedType: (selectedType: GRAPH_TYPE) => set({ selectedType }),
 }));
