+++ apps/www/src/page/bookmark-page.tsx
@@ -4,6 +4,9 @@ import { Usable, use, useState } from "react";
 
 import Graph from "@/components/bookmarks/graph";
 import GraphDetail from "@/components/bookmarks/graph/graph-detail";
+import Planet from "@/components/bookmarks/graph/planet";
+import { GRAPH_THEME } from "@/constants/bookmark";
+import { useBookmarkStore } from "@/lib/zustand/bookmark";
 import { BaseResponseDTO } from "@/models";
 import { AllStarDTO } from "@/models/star";
 
@@ -18,6 +21,8 @@ const BookmarkPage = ({ promisedData }: BookmarkPageProps) => {
   });
 
   const data = use<BaseResponseDTO<AllStarDTO>>(promisedData);
+
+  const bookmarkStore = useBookmarkStore();
 
   const onOpen = (id: string) => {
     setDetail({
@@ -35,7 +40,11 @@ const BookmarkPage = ({ promisedData }: BookmarkPageProps) => {
 
   return (
     <>
-      <Graph onOpen={onOpen} data={data?.result} />
+      {bookmarkStore.selectedTheme === GRAPH_THEME.GRAPH ? (
+        <Graph onOpen={onOpen} data={data?.result} />
+      ) : (
+        <Planet onOpen={onOpen} data={data?.result} />
+      )}
       <GraphDetail open={detail.open} id={detail.id} onClose={onClose} />
     </>
   );
